Moved EndpointScaling.hpp to separate file.
