+++ opm/parser/eclipse/EclipseState/EndpointScaling.cpp
@@ -1,5 +1,5 @@
 #include <opm/parser/eclipse/Deck/Deck.hpp>
-#include <opm/parser/eclipse/EclipseState/Runspec.hpp>
+#include <opm/parser/eclipse/EclipseState/EndpointScaling.hpp>
 #include <opm/parser/eclipse/Utility/String.hpp>
 
 namespace Opm {
