+++ opm/parser/eclipse/EclipseState/Runspec.hpp
@@ -19,15 +19,15 @@
 #ifndef OPM_RUNSPEC_HPP
 #define OPM_RUNSPEC_HPP
 
-#include <bitset>
 #include <iosfwd>
 #include <string>
 
 #include <opm/parser/eclipse/EclipseState/Tables/Tabdims.hpp>
+#include <opm/parser/eclipse/EclipseState/EndpointScaling.hpp>
 
 namespace Opm {
-
 class Deck;
+
 
 enum class Phase {
     OIL     = 0,
@@ -51,35 +51,9 @@ class Phases {
         std::bitset< 4 > bits;
 };
 
-class EndpointScaling {
-    public:
-        EndpointScaling() noexcept = default;
-        explicit EndpointScaling( const Deck& );
-
-        /* true if endpoint scaling is enabled, otherwise false */
-        operator bool() const noexcept;
-
-        bool directional() const noexcept;
-        bool nondirectional() const noexcept;
-        bool reversible() const noexcept;
-        bool irreversible() const noexcept;
-        bool twopoint() const noexcept;
-        bool threepoint() const noexcept;
-
-    private:
-        enum class option {
-            any         = 0,
-            directional = 1,
-            reversible  = 2,
-            threepoint  = 3,
-        };
-
-        using ue = std::underlying_type< option >::type;
-        std::bitset< 4 > options;
-};
 
 class Runspec {
-    public:
+   public:
         explicit Runspec( const Deck& );
 
         const Phases& phases() const noexcept;
